requirements.txt
fastapi
uvicorn
sqlalchemy
alembic
psycopg2-binary
python-dotenv
passlib[bcrypt]
pyjwt
pydantic[email]
python-jose[cryptography]
requests
authlib
oauthlib
httpx
starlette
itsdangerous
